+++ rewatch/src/cli.rs
@@ -194,21 +194,6 @@ pub struct AfterBuildArg {
     pub after_build: Option<String>,
 }
 
-#[derive(Args, Debug, Clone, Copy)]
-pub struct CreateSourceDirsArg {
-    /// Deprecated: source_dirs.json is now always created.
-    #[arg(short, long, num_args = 0..=1, default_missing_value = "true", hide = true)]
-    pub create_sourcedirs: Option<bool>,
-}
-
-#[derive(Args, Debug, Clone, Copy)]
-pub struct DevArg {
-    /// Deprecated: Build development dependencies
-    /// This is the flag no longer does anything and will be removed in future versions.
-    #[arg(long, default_value_t = false, num_args = 0..=1, hide = true)]
-    pub dev: bool,
-}
-
 #[derive(Args, Debug, Clone)]
 pub struct WarnErrorArg {
     /// Override warning configuration from rescript.json.
@@ -229,21 +214,11 @@ pub struct BuildArgs {
     pub after_build: AfterBuildArg,
 
     #[command(flatten)]
-    pub create_sourcedirs: CreateSourceDirsArg,
-
-    #[command(flatten)]
-    pub dev: DevArg,
-
-    #[command(flatten)]
     pub warn_error: WarnErrorArg,
 
     /// Disable output timing
     #[arg(short, long, default_value_t = false, num_args = 0..=1)]
     pub no_timing: bool,
-
-    /// Watch mode (deprecated, use `rescript watch` instead)
-    #[arg(short, default_value_t = false, num_args = 0..=1, hide = true)]
-    pub watch: bool,
 }
 
 #[cfg(test)]
@@ -396,12 +371,6 @@ pub struct WatchArgs {
     pub after_build: AfterBuildArg,
 
     #[command(flatten)]
-    pub create_sourcedirs: CreateSourceDirsArg,
-
-    #[command(flatten)]
-    pub dev: DevArg,
-
-    #[command(flatten)]
     pub warn_error: WarnErrorArg,
 }
 
@@ -411,8 +380,6 @@ impl From<BuildArgs> for WatchArgs {
             folder: build_args.folder,
             filter: build_args.filter,
             after_build: build_args.after_build,
-            create_sourcedirs: build_args.create_sourcedirs,
-            dev: build_args.dev,
             warn_error: build_args.warn_error,
         }
     }
@@ -428,9 +395,6 @@ pub enum Command {
     Clean {
         #[command(flatten)]
         folder: FolderArg,
-
-        #[command(flatten)]
-        dev: DevArg,
     },
     /// Format ReScript files.
     Format {
@@ -451,9 +415,6 @@ pub enum Command {
         /// Files to format. If no files are provided, all files are formatted.
         #[arg(group = "format_input_mode")]
         files: Vec<String>,
-
-        #[command(flatten)]
-        dev: DevArg,
     },
     /// Print the compiler arguments for a ReScript source file.
     CompilerArgs {
@@ -484,21 +445,6 @@ impl Deref for AfterBuildArg {
 
     fn deref(&self) -> &Self::Target {
         &self.after_build
-    }
-}
-
-impl CreateSourceDirsArg {
-    /// Returns true if the flag was explicitly passed on the command line.
-    pub fn was_explicitly_set(&self) -> bool {
-        self.create_sourcedirs.is_some()
-    }
-}
-
-impl Deref for DevArg {
-    type Target = bool;
-
-    fn deref(&self) -> &Self::Target {
-        &self.dev
     }
 }